Add Cootue curses square font as the default font in the web interface.
There probably will be more fonts later but this is better than the default
that the web browser gives you.

I asked and received permission to bundle this with Dfterm3 on 2013-04-08.  The
following conversation is from the bay12 forums. I attached it here so that the
permission is documented somewhere.

--------------------------------------------------------------------------
> Adeon wrote:
>
> Hi!
>
> Are you the person who made the Cootue curses square font? This one:
> http://dffd.wimbli.com/file.php?id=1854
>
> I've played around with TTF fonts in my upcoming Dwarf Fortress web-interface
> playing software and might want to use that font as one of the default fonts
> you can select. However, I cannot just do that without knowing who owns the
> copyright on this and if they are willing to let me bundle it.
>
> If you own the copyright on this font, would you be willing to let me bundle
> it with Dfterm3 as long as I credit you and won't commercialize Dfterm3?

Cooz wrote:

Hi,
yes, I've created it and I'm glad it might be useful for something
Sure you can bundle it with Dfterm3.

Cheers and good luck with your project.