Add some server side-code for administrator panel.
I have implemented the logging-in functionality but the admin page itself is
empty.

Administrators are tracked by cookies that store the session id and the cookies
expire after 30 minutes or if the browser is closed.

Unfortunately there is no option of using SSL/TLS connections. However, the
administrator panel is served only on 127.0.0.1 so it cannot be directly
accessed from outside. However, in the future it may be useful to remotely
access the panel and it must have encryption if we'll do so.

Some parts of Happstack are now a dependency as I used that to provide the
admin interface. Perhaps I'll serve the other files (i.e. the playing
interface) through Happstack as well in the future.