Don't hardcode WebSocket ports.
Until now, the port 8000 was hard-coded to playing.html (the playing interface)
as the WebSocket port. Now, the command line options are taken into account and
passed on to the client browser so that it knows which ports it should really
connect into.

If more than one port is specified, a port is selected at random.