Rewrite lots of...stuff.
If it's not obvious from the changeset, this is almost a rewrite of Dfterm3. As
of this commit, the software is broken but compiles. The only reason I make
this commit is that I need at least some git history and not just commit a huge
one commit at the end.

I will add bits of old code back to this repository as I make things work
again.

The previous model to work with games was not very flexible. It was difficult
to extend and in particular, GitHub issue #4 would be difficult to properly
fix.