Modernize dependencies. BREAKS WebSocket.
I checked all the dependencies and took the newest ones. The major upgrade is
the websockets package upgrade from 0.7 to 0.8 which unfortunately has a
completely different API. As a result, as of this commit, the WebSocket
interface is broken.

That would be a good time to fix the protocol too when I repair it.

I checked all the dependencies and made the package require the latest versions
of them as of now and allowing upgrades until the next major version upgrade.
(see changes in the .cabal file).

Added Dfterm3.Prelude and a whole bunch of default extensions to the cabal
file. This may help to curb massive import lists in some modules.