Abstract away the user session semantics.
It used to be that the protocol semantics are tied to the WebSocket
implementation. Now I have made it so that the code that handles WebSocket
connection and code that handles the logic of session handling etc. are
separate.

This makes the code cleaner and easier to manage. Also, it makes it easier to
later add alternative frontends, should we ever want to do that for some
reason.