Finish up the protocol code.
Now it's complete on the server and client side. However, as of this commit,
none of it is tested so it probably won't work right away.

An interesting new file is web/src/dfterm3_protocol.js. I have implemented a
node.js module that implements the protocol so that it's easy and convenient to
use from JavaScript and is completely agnostic to how the Dwarf Fortress data
is to be handled. With browserify, we can use it in the browser.